requirements.txt
aiogram>=3.12,<4.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
python-dotenv>=1.0
